# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoundSelect",
    "description": "\nSoundSelect's back end. Every screen talks to it through these calls.\n\nSongs come back as the **Song result** (see the `Song` schema): concert pitch and written for\nthe instrument, with the key's pentatonic first. Errors come as\n`{\"detail\": {\"code\": ..., \"message\": ...}}`, with a message that can be shown as it is.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/batches/{batch_id}": {
      "get": {
        "tags": [
          "imports"
        ],
        "summary": "Get Batch",
        "description": "A batch's progress: each song's status and current step.",
        "operationId": "get_batch_api_v1_batches__batch_id__get",
        "parameters": [
          {
            "name": "batch_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Batch Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchInfo"
                }
              }
            }
          },
          "400": {
            "description": "The request can't be used; see the message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/songs/{song_id}": {
      "get": {
        "tags": [
          "songs"
        ],
        "summary": "Get Song",
        "description": "The full Song result, written for the instrument, with a table of note names.",
        "operationId": "get_song_api_v1_songs__song_id__get",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Song Id"
            }
          },
          {
            "name": "instrument",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Instrument to write for; your default when left out.",
              "title": "Instrument"
            },
            "description": "Instrument to write for; your default when left out."
          },
          {
            "name": "names",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "letters",
                    "russian",
                    "solfege",
                    "german",
                    "both",
                    "none"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Note names to add (or show on the staff); your default when left out.",
              "title": "Names"
            },
            "description": "Note names to add (or show on the staff); your default when left out."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          },
          "400": {
            "description": "The request can't be used; see the message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "songs"
        ],
        "summary": "Patch Song",
        "description": "Save corrections (the key, the capo, a chord, the title) and get the song back once the\nsteps they touch have run again. Fields left out stay as they are; null clears one. An\nempty body makes the song again with the newest version of every step.",
        "operationId": "patch_song_api_v1_songs__song_id__patch",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Song Id"
            }
          },
          {
            "name": "instrument",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Instrument to write for; your default when left out.",
              "title": "Instrument"
            },
            "description": "Instrument to write for; your default when left out."
          },
          {
            "name": "names",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "letters",
                    "russian",
                    "solfege",
                    "german",
                    "both",
                    "none"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Note names to add (or show on the staff); your default when left out.",
              "title": "Names"
            },
            "description": "Note names to add (or show on the staff); your default when left out."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SongPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          },
          "400": {
            "description": "The request can't be used; see the message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "songs"
        ],
        "summary": "Delete Song",
        "operationId": "delete_song_api_v1_songs__song_id__delete",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Song Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "description": "The request can't be used; see the message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings": {
      "get": {
        "tags": [
          "settings"
        ],
        "summary": "Get Settings",
        "description": "Your defaults: instrument, note names, comfortable range, photo reader.",
        "operationId": "get_settings_api_v1_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "settings"
        ],
        "summary": "Put Settings",
        "operationId": "put_settings_api_v1_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Settings"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatchInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "status": {
            "type": "string",
            "enum": [
              "working",
              "done"
            ],
            "title": "Status",
            "description": "'done' once every job has finished, even when some failed."
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "done": {
            "type": "integer",
            "title": "Done"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "options": {
            "$ref": "#/components/schemas/ImportOptions"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobInfo"
            },
            "type": "array",
            "title": "Jobs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "created_at",
          "status",
          "total",
          "done",
          "failed",
          "options",
          "jobs"
        ],
        "title": "BatchInfo",
        "description": "One import: a job for each song, filled in song by song."
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportOptions": {
        "properties": {
          "instrument": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrument",
            "description": "Instrument to write for; empty means your default."
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key",
            "description": "The concert key, when you know it: 'Bb', 'Gm'."
          },
          "capo": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Capo"
          },
          "groups": {
            "anyOf": [
              {
                "items": {
                  "items": {
                    "type": "integer"
                  },
                  "type": "array"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Groups",
            "description": "Files that together make one song, as lists of file positions (from 0); for photos of a song that runs over two pages. Read from the photo phase on."
          },
          "link_mode": {
            "type": "string",
            "enum": [
              "auto",
              "sound",
              "sheet_music"
            ],
            "title": "Link Mode",
            "description": "Whether a link is read for its sound (song tool) or the sheet music it shows. Read from the phases that add links.",
            "default": "auto"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ImportOptions",
        "description": "Options for one import. Title, artist, key and capo apply when it holds one song."
      },
...
      "JobInfo": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "position": {
            "type": "integer",
            "title": "Position",
            "description": "Place in the batch, from 0."
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "File name, link, or the first line of pasted text."
          },
          "kind": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "chord_sheet",
                  "sheet_music",
                  "song"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind",
            "description": "The tool that reads it; null when none can yet."
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/InputRef"
            },
            "type": "array",
            "title": "Inputs"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "done",
              "failed"
            ],
            "title": "Status"
          },
          "step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step",
            "description": "The step running now, e.g. 'key'."
          },
          "step_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Step Label",
            "description": "That step in words: 'Finding the key'."
          },
          "step_number": {
            "type": "integer",
            "title": "Step Number",
            "description": "The running step's number, from 1.",
            "default": 0
          },
          "step_count": {
            "type": "integer",
            "title": "Step Count",
            "default": 0
          },
          "progress": {
            "type": "number",
            "title": "Progress",
            "description": "0 to 1.",
            "default": 0.0
          },
          "song_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Song Id"
          },
          "reused": {
            "type": "boolean",
            "title": "Reused",
            "description": "The song was already in the library.",
            "default": false
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error",
            "description": "Why it failed, in words for the player."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "rev": {
            "type": "integer",
            "title": "Rev",
            "description": "Goes up with every change; the event streams use it."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "batch_id",
          "position",
          "name",
          "kind",
          "inputs",
          "status",
          "created_at",
          "rev"
        ],
        "title": "JobInfo",
        "description": "One song being made: which step it is on, and the song once it's ready."
      },
...
      "Settings": {
        "properties": {
          "instrument": {
            "type": "string",
            "title": "Instrument",
            "default": "alto_sax"
          },
          "comfortable_low": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comfortable Low",
            "description": "Written note; instrument default."
          },
          "comfortable_high": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comfortable High",
            "description": "Written note; instrument default."
          },
          "names": {
            "type": "string",
            "enum": [
              "letters",
              "russian",
              "solfege",
              "german",
              "both",
              "none"
            ],
            "title": "Names",
            "default": "russian"
          },
          "photo_reader": {
            "type": "string",
            "enum": [
              "local",
              "ai"
            ],
            "title": "Photo Reader",
            "default": "local"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Settings"
      },
      "Song": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "song",
            "title": "Kind",
            "default": "song"
          },
          "schema_version": {
            "type": "integer",
            "title": "Schema Version",
            "default": 1
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "identity": {
            "$ref": "#/components/schemas/Identity"
          },
          "key": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/KeyResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "view": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InstrumentView"
              },
              {
                "type": "null"
              }
            ]
          },
          "form": {
            "items": {
              "$ref": "#/components/schemas/Section"
            },
            "type": "array",
            "title": "Form"
          },
          "chords": {
            "items": {
              "$ref": "#/components/schemas/ChordInfo"
            },
            "type": "array",
            "title": "Chords"
          },
          "scales": {
            "items": {
              "$ref": "#/components/schemas/ScaleInfo"
            },
            "type": "array",
            "title": "Scales",
            "description": "Whole-song scales; the first, role 'headline', is the key's pentatonic."
          },
          "melody": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Melody"
              },
              {
                "type": "null"
              }
            ]
          },
          "timing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Timing"
              },
              {
                "type": "null"
              }
            ]
          },
          "corrections": {
            "$ref": "#/components/schemas/Corrections"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/Notice"
            },
            "type": "array",
            "title": "Notes",
            "description": "Notes to the player."
          },
          "source_pages": {
            "items": {
              "$ref": "#/components/schemas/SourcePage"
            },
            "type": "array",
            "title": "Source Pages"
          },
          "versions": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Versions",
            "description": "Version of each step that produced this result."
          },
          "names": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Names",
            "description": "Display names for every pitch in the song, keyed by its text without octave ('Bb' -> 'си♭'); present when names were asked for."
          },
          "name_system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name System"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "identity"
        ],
        "title": "Song"
      },
...
      "SongPatch": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key",
            "description": "The right concert key: 'Bb', 'Gm', 'ля минор'."
          },
          "capo": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Capo"
          },
          "chords": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ChordFix"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chords",
            "description": "The whole list of chord fixes (send the list with yours added)."
          },
          "melody_octave": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 3.0,
                "minimum": -3.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Melody Octave"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "SongPatch",
        "description": "Corrections to save. Fields left out stay as they are; null clears one."
      },
...
    }
  }
}